requirements.txt
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
click==8.1.2
dnspython==2.2.1
email-validator==1.1.3
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pycparser==2.21
six==1.16.0
SQLAlchemy==1.4.35
sqlparse==0.4.2
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0
python-dotenv
s3transfer==0.6.0
boto3==1.26.9
botocore==1.29.9
Pillow==9.3.0